# API - updated/openapi.yaml
openapi: 3.0.0
info:
  title: WDE - Assignment 3.1 -  updated
  description: Recipes website API
  contact:
    name: yarden curiel & aviv amsellem.
    email: curiey@post.bgu.ac.il
  version: 1.1.1
servers:
- url: http://localhost:3000/
...
paths:
...
  /getRandomRecipes/explore:
    get:
      tags:
      - site
      summary: home page with recipes.
      description: home page. receive 2 arrays. one is 3 random recipes and second
        is 3 last viewed recipes. for guest the second array will be empty.
      responses:
        "200":
          description: OK. recipes on the way
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "204":
          description: Bad, No Content was found.
        "400":
          description: Bad Request.
  /login:
    post:
      tags:
      - user
      summary: login method.
      description: Login and set session id.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        "200":
          description: Login Successfully. The session ID is returned in a cookie
            named `JSESSIONID`. You need to include this cookie in subsequent requests.
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: JSESSIONID=abcde12345
        "401":
          description: Autorization failed, Username or Password incorrect.
  /register:
    post:
      tags:
      - user
      summary: Registration method.
      description: Registration page for a new user.
      requestBody:
        description: Registration page for a new user.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_1'
        required: true
      responses:
        "200":
          description: OK. registration complete successfully
        "400":
          description: Bad Request. Not all reqired argument was given.
        "401":
          description: User already logged in.
        "409":
          description: Username is already taken.
  /logout:
    get:
      tags:
      - user
      summary: User log out from website.
      description: User Request to logout from the website.
      responses:
        "200":
          description: logout succeeded.
        "401":
          description: must logged in before logging out.
      security:
      - cookieAuth: []
  /viewOurFamilyRecipes:
    get:
      tags:
      - site
      summary: Get developers recipes.
      description: Get all the web developers family recipes.
      responses:
        "200":
          description: OK. recipes on the way
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PreviewRecipe'
        "204":
          description: Bad, No Content was found.
        "400":
          description: Bad Request.
...
components:
  schemas:
...
    BasePreviewPecipe:
      required:
      - aggregateLikes
      - glutenFree
      - image
      - instructions
      - isBeenWatched
      - isFavorite
      - readyInMinutes
      - title
      - vegan
      - vegetarian
      properties:
        title:
          type: string
          example: gahnun
        readyInMinutes:
          type: integer
          example: 7
        aggregateLikes:
          type: integer
          example: 453
        vegan:
          type: boolean
          example: true
        vegetarian:
          type: boolean
          example: true
        glutenFree:
          type: boolean
          example: true
        isBeenWatched:
          type: boolean
          example: true
        isFavorite:
          type: boolean
          example: true
        image:
          type: string
          example: gahnun.jpg
      description: Base preview of recipe. containing some of the recipe information
        without id field  (use for upload a new recipe).
    PreviewRecipe:
      description: Preview of recipe. conatining some or recipe information with id
        field.
      allOf:
      - $ref: '#/components/schemas/BasePreviewPecipe'
      - required:
        - id
        type: object
        properties:
          id:
            type: integer
            example: 453
...
    inline_response_200_1:
      type: object
      properties:
        RandomRecipes:
          maxItems: 3
          minItems: 3
          type: array
          items:
            $ref: '#/components/schemas/PreviewRecipe'
    body:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: Containing letter only.
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: Containing at least one letter and at least one numbers.
      example:
        username: admin
        password: a123456
    body_1:
      type: object
      properties:
        username:
          type: string
          example: Johnsmith1
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        country:
          type: string
          example: Israel
        password:
          type: string
          example: password
        email:
          type: string
          example: johnsmith@gmail.com
        image:
          type: string
          example: https://cloudinary.com/image.jpg
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: session
      in: cookie